changed all doubles used throughout the code to BigDecimal, for more precise decimal arithmetic.
added BigNumber to JS, which is like BigDecimal for JS, for basic front-end arithmetic

+++ junopoker/src/main/java/com/v1/junopoker/callback/TableCallback.java
@@ -3,19 +3,20 @@ package com.v1.junopoker.callback;
 import com.v1.junopoker.model.Card;
 import com.v1.junopoker.model.Player;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.HashMap;
 
 public interface TableCallback {
     void onButtonSet(int buttonIndex);
 
-    void onPotInit(int sbIndex, int bbIndex, double sbAmount, double bbAmount, double potSize);
+    void onPotInit(int sbIndex, int bbIndex, BigDecimal sbAmount, BigDecimal bbAmount, BigDecimal potSize);
 
     void onHoleCardsDealt(String username, int seat, Card[] holeCards);
 
-    void onPreFlopAction(Player player, int seat, double currentBet, double potSize, double minBet);
+    void onPreFlopAction(Player player, int seat, BigDecimal currentBet, BigDecimal potSize, BigDecimal minBet);
 
-    void onEndPlayerAction(char action, String username, int seatIndex, double betAmount, double stackSize, double potSize, double currentStreetPotSize, boolean isPreFlop);
+    void onEndPlayerAction(char action, String username, int seatIndex, BigDecimal betAmount, BigDecimal stackSize, BigDecimal potSize, BigDecimal currentStreetPotSize, boolean isPreFlop);
 
     void onShowdown(HashMap<Integer, Player> indexAndPlayer);
     void onCompleteHand(HashMap<Integer, Player> indexAndWinner);
