+++ junopoker/src/main/java/com/v1/junopoker/dto/PlayerActionRequest.java
@@ -3,13 +3,15 @@ package com.v1.junopoker.dto;
 import com.v1.junopoker.model.Player;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 
 public class PlayerActionRequest {
     private RequestType type;
     private Player player;
     private int seat;
-    private double currentBet;
-    private double potSize;
-    private double minBet;
+    private BigDecimal currentBet;
+    private BigDecimal potSize;
+    private BigDecimal minBet;
 }
