+++ junopoker/src/main/java/com/v1/junopoker/dto/PlayerActionResponse.java
@@ -2,11 +2,13 @@ package com.v1.junopoker.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class PlayerActionResponse {
     private char action;
-    private double betAmount;
+    private BigDecimal betAmount;
     private int seat;
-    private double stackSize;
-    private double potSize;
+    private BigDecimal stackSize;
+    private BigDecimal potSize;
 }
